Remove circular dep
